move specific nameversion implementations to their runners
