+++ internal/runners/platforms/platforms.go
@@ -13,6 +13,18 @@ import (
 	"github.com/ActiveState/cli/pkg/platform/model"
 )
 
+type PlatformVersion struct {
+	captain.NameVersion
+}
+
+func (pv *PlatformVersion) Set(arg string) error {
+	err := pv.NameVersion.Set(arg)
+	if err != nil {
+		return locale.NewError("err_platform_format", "The platform and version provided is not formatting correctly, must be in the form of <platform>@<version>")
+	}
+	return nil
+}
+
 // Platform represents the output data of a platform.
 type Platform struct {
 	Name     string `json:"name"`
